[discussion] Retries email sending
* reties
* applies backoff
* is tolerant to restarts and deployment of multiple instances

Instead of MongoDB, this uses RabbitMQ with the delayed exchange plugin.
The idea is, that failed messages are resent to the message queue (with a delay) up to 4 times. This acts as external storage and makes the system tolerant to restarts.
Since RabbitMQ is poll-based, this works with any number of instances.